Moved common jsons and case classes to fixture
