+++ test/v1/controllers/GetIncomeTaxCalcControllerSpec.scala
@@ -16,10 +16,11 @@
 
 package v1.controllers
 
-import play.api.libs.json.{JsValue, Json}
+import play.api.libs.json.Json
 import play.api.mvc.Result
 import uk.gov.hmrc.domain.Nino
 import uk.gov.hmrc.http.HeaderCarrier
+import v1.fixtures.GetIncomeTaxCalcFixture
 import v1.handling.RequestDefn
 import v1.mocks.requestParsers.MockGetCalculationParser
 import v1.mocks.services.{MockEnrolmentsAuthService, MockMtdIdLookupService, MockStandardService}
@@ -27,7 +28,6 @@ import v1.models.errors._
 import v1.models.outcomes.ResponseWrapper
 import v1.models.request.{GetCalculationRawData, GetCalculationRequest}
 import v1.models.response.CalculationWrapperOrError
-import v1.models.response.getIncomeTaxCalc.{CalculationDetail, _}
 
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
@@ -58,36 +58,6 @@ class GetIncomeTaxCalcControllerSpec
   private val calcId        = "someCalcId"
   private val correlationId = "X-123"
 
-  val incomeTaxSummary = IncomeTaxSummary(100.25, None, None)
-  val nicSummary = NicSummary(Some(200.25), None, None)
-  val calculationSummary = CalculationSummary(incomeTaxSummary, Some(nicSummary), Some(300.25), Some(400.25), 500.25, "UK")
-
-  val incomeTaxDetail = IncomeTaxDetail(Some(IncomeTypeBreakdown(100.25, 200.25, None)), None, None, None)
-  val nicDetail = NicDetail(Some(Class2NicDetail(Some(300.25), None, None, None, underSmallProfitThreshold = true,
-    actualClass2Nic = Some(false))), None)
-  val taxDeductedAtSource = TaxDeductedAtSource(Some(400.25), None)
-  val calculationDetail: CalculationDetail = CalculationDetail(incomeTaxDetail, Some(nicDetail), Some(taxDeductedAtSource))
-
-  val response = CalculationWrapperOrError.CalculationWrapper(GetIncomeTaxCalcResponse(summary = calculationSummary,
-     detail = calculationDetail))
-
-  val responseBody: JsValue = Json.parse(s"""{
-                                            |  "summary": {
-                                            |     "incomeTax" : ${Json.toJson(incomeTaxSummary).toString()},
-                                            |     "nics" : ${Json.toJson(nicSummary).toString()},
-                                            |     "totalIncomeTaxNicsCharged" : 300.25,
-                                            |     "totalTaxDeducted" : 400.25,
-                                            |     "totalIncomeTaxAndNicsDue" : 500.25,
-                                            |     "taxRegime" : "UK"
-                                            |  },
-                                            |  "detail": {
-                                            |     "incomeTax" : ${Json.toJson(incomeTaxDetail).toString()},
-                                            |     "nics" : ${Json.toJson(nicDetail).toString()},
-                                            |     "taxDeductedAtSource" : ${Json.toJson(taxDeductedAtSource).toString()}
-                                            |  }
-                                            |}
-                                            |""".stripMargin)
-
   private val rawData     = GetCalculationRawData(nino, calcId)
   private val requestData = GetCalculationRequest(Nino(nino), calcId)
 
@@ -102,12 +72,12 @@ class GetIncomeTaxCalcControllerSpec
 
         MockStandardService
           .doService(RequestDefn.Get(uri), OK)
-          .returns(Future.successful(Right(ResponseWrapper(correlationId, response))))
+          .returns(Future.successful(Right(ResponseWrapper(correlationId, GetIncomeTaxCalcFixture.getIncomeTaxCalcResponseObj))))
 
         val result: Future[Result] = controller.getIncomeTaxCalc(nino, calcId)(fakeGetRequest(uri))
 
         status(result) shouldBe OK
-        contentAsJson(result) shouldBe responseBody
+        contentAsJson(result) shouldBe GetIncomeTaxCalcFixture.successOutputToVendor
         header("X-CorrelationId", result) shouldBe Some(correlationId)
       }
     }
